extract MetadataFilterChips state and logic into useMetadataFilterChips hook
